OSINT_X-IG/osint-profile-linker/requirements.txt
streamlit>=1.25
instaloader>=4.9
requests>=2.28
sentence-transformers>=2.2.2
numpy
pandas
Pillow
imagehash
python-dotenv
scikit-learn
tqdm
